refactor / add test for schema
