+++ emmet-core/emmet/core/math.py
@@ -23,3 +23,18 @@ ListVector3D.__doc__ = "Real space vector as list"  # type: ignore
 
 ListMatrix3D = List[ListVector3D]
 ListMatrix3D.__doc__ = "Real space Matrix as list"  # type: ignore
+
+VOIGT_INDICES = [(0, 0), (1, 1), (2, 2), (1, 2), (0, 2), (0, 1)]
+
+def matrix_3x3_to_voigt(matrix : Matrix3D ) -> Vector6D:
+    """Convert a 3x3 symmetric matrix to its Voigt vector representation.
+    
+    Parameters
+    -----------
+    matrix : Matrix3D
+    
+    Returns
+    -----------
+    Vector6D
+    """
+    return [matrix[idx[0]][idx[1]] for idx in VOIGT_INDICES]
