+++ emmet-core/emmet/core/ml.py
@@ -1,9 +1,23 @@
-from typing import TYPE_CHECKING, Dict, List, Optional, Union
+"""Define schemas for ML calculations and training data."""
 
-from matcalc import ElasticityCalc, EOSCalc, PESCalculator, PhononCalc, RelaxCalc
-from pydantic import Field, field_validator
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+import numpy as np
+
+try:
+    from matcalc import ElasticityCalc, EOSCalc, PESCalculator, PhononCalc, RelaxCalc
+except ImportError:
+    ElasticityCalc = None
+    EOSCalc = None
+    PESCalculator = None
+    PhononCalc = None
+    RelaxCalc = None
+
+from pydantic import Field, BaseModel, field_validator, model_validator
 from pymatgen.analysis.elasticity import ElasticTensor
-from pymatgen.core import Structure
+from pymatgen.core import Composition, Element, Structure
 
 from emmet.core.elasticity import (
     BulkModulus,
@@ -11,10 +25,19 @@ from emmet.core.elasticity import (
     ElasticTensorDoc,
     ShearModulus,
 )
+from emmet.core.math import Vector3D, Matrix3D, Vector6D, matrix_3x3_to_voigt
+from emmet.core.mpid import MPID
+from emmet.core.structure import StructureMetadata
+from emmet.core.vasp.calc_types import RunType as VaspRunType
+from emmet.core.tasks import TaskDoc
 
 if TYPE_CHECKING:
-    from ase.calculators.calculator import Calculator
+    from typing_extensions import Self
 
+    try:
+        from ase.calculators.calculator import Calculator as AseCalculator
+    except ImportError:
+        AseCalculator = None    
 
 class MLDoc(ElasticityDoc):
     """Document model for matcalc-generated material properties from machine learning
@@ -51,47 +74,47 @@ class MLDoc(ElasticityDoc):
 
     # metadata
     structure: Structure = Field(description="Original structure")
-    matcalc_version: Optional[str] = Field(
+    matcalc_version: str | None = Field(
         None, description="Version of matcalc used to generate this document"
     )
-    model: Optional[str] = Field(
+    model: str | None = Field(
         None, description="Name of model used as ML potential."
     )
-    version: Optional[str] = Field(
+    version: str | None = Field(
         None, description="Version of model used as ML potential"
     )
 
     # relaxation attributes
-    final_structure: Optional[Structure] = Field(
+    final_structure: Structure | None = Field(
         None, description="ML-potential-relaxed structure"
     )
-    energy: Optional[float] = Field(None, description="Final energy in eV")
-    volume: Optional[float] = Field(None, description="Final volume in Angstrom^3")
-    a: Optional[float] = Field(None, description="Lattice length a in Angstrom")
-    b: Optional[float] = Field(None, description="Lattice length b in Angstrom")
-    c: Optional[float] = Field(None, description="Lattice length c in Angstrom")
-    alpha: Optional[float] = Field(None, description="Lattice angle alpha in degrees")
-    beta: Optional[float] = Field(None, description="Lattice angle beta in degrees")
-    gamma: Optional[float] = Field(None, description="Lattice angle gamma in degrees")
+    energy: float | None = Field(None, description="Final energy in eV")
+    volume: float | None = Field(None, description="Final volume in Angstrom^3")
+    a: float | None = Field(None, description="Lattice length a in Angstrom")
+    b: float | None = Field(None, description="Lattice length b in Angstrom")
+    c: float | None = Field(None, description="Lattice length c in Angstrom")
+    alpha: float | None = Field(None, description="Lattice angle alpha in degrees")
+    beta: float | None = Field(None, description="Lattice angle beta in degrees")
+    gamma: float | None = Field(None, description="Lattice angle gamma in degrees")
 
     # equation of state attributes
-    eos: Optional[Dict[str, List[float]]] = Field(
+    eos: dict[str, list[float]] | None = Field(
         None, description="dict with keys energies and volumes"
     )
-    bulk_modulus_bm: Optional[float] = Field(None, description="bm.b0_GPa")
+    bulk_modulus_bm: float | None = Field(None, description="bm.b0_GPa")
 
     # phonons attributes
-    temperatures: Optional[List[float]] = Field(
+    temperatures: list[float] | None = Field(
         None, description="list of temperatures"
     )
-    free_energy: Optional[List[float]] = Field(
+    free_energy: list[float] | None = Field(
         None,
         description="list of Helmholtz free energies at corresponding temperatures",
     )
-    entropy: Optional[List[float]] = Field(
+    entropy: list[float] | None = Field(
         None, description="list of entropies at corresponding temperatures in eV/K"
     )
-    heat_capacity: Optional[List[float]] = Field(
+    heat_capacity: list[float] | None = Field(
         None,
         description="list of heat capacities at constant volume at corresponding "
         "temperatures in eV/K",
@@ -129,8 +152,8 @@ class MLDoc(ElasticityDoc):
         cls,
         structure,
         material_id,
-        calculator: Union[str, "Calculator"],
-        prop_kwargs: Optional[dict] = None,
+        calculator: str | AseCalculator,
+        prop_kwargs: dict | None = None,
         **kwargs,
     ) -> None:
         """
@@ -168,3 +191,219 @@ class MLDoc(ElasticityDoc):
         super().__init__(
             structure=structure, material_id=material_id, **results, **kwargs
         )
+
+class MLTrainDoc(StructureMetadata):
+    """Generic schema for ML training data."""
+
+    structure: Structure | None = Field(None, description="Structure for this entry.")
+
+    energy: float | None = Field(
+        None, description="The total energy associated with this structure."
+    )
+
+    forces: list[Vector3D] | None = Field(
+        None,
+        description="The interatomic forces corresponding to each site in the structure.",
+    )
+
+    abs_forces: list[float] | None = Field(
+        None, description="The magnitude of the interatomic force on each site."
+    )
+
+    stress: Vector6D | None = Field(
+        None,
+        description="The components of the symmetric stress tensor in Voigt notation (xx, yy, zz, yz, xz, xy).",
+    )
+
+    stress_matrix : Matrix3D | None = Field(
+        None,
+        description = "The 3x3 stress tensor. Use this if the tensor is unphysically non-symmetric."
+    )
+
+    bandgap: float | None = Field(None, description="The final DFT bandgap.")
+
+    elements: list[Element] | None = Field(
+        None,
+        description="List of unique elements in the material sorted alphabetically.",
+    )
+
+    composition: Composition | None = Field(
+        None, description="Full composition for the material."
+    )
+
+    composition_reduced: Composition | None = Field(
+        None,
+        title="Reduced Composition",
+        description="Simplified representation of the composition.",
+    )
+
+    functional: VaspRunType | None = Field(
+        None, description="The approximate functional used to generate this entry."
+    )
+
+    bader_charges: list[float] | None = Field(
+        None, description="Bader charges on each site of the structure."
+    )
+    bader_magmoms: list[float] | None = Field(
+        None,
+        description="Bader on-site magnetic moments for each site of the structure.",
+    )
+
+    @model_validator(mode="after")
+    def set_abs_forces(
+        self,
+    ) -> Self:
+        """Ensure the abs_forces are set if the vector-valued forces are."""
+        if self.forces is not None and self.abs_forces is None:
+            self.abs_forces = [np.linalg.norm(f) for f in self.forces]
+        return self
+
+    @classmethod
+    def from_structure(
+        cls,
+        meta_structure: Structure,
+        fields: list[str] | None = None,
+        **kwargs,
+    ) -> Self:
+        """
+        Create an ML training document from a structure and fields.
+        
+        This method mostly exists to ensure that the structure field is
+        set because `meta_structure` does not populate it automatically.
+
+        Parameters
+        -----------
+        meta_structure : Structure
+        fields : list of str or None
+            Additional fields in the document to populate
+        **kwargs
+            Any other fields / constructor kwargs
+        """
+        return super().from_structure(
+            meta_structure=meta_structure,
+            fields=fields,
+            structure=meta_structure,
+            **kwargs,
+        )
+
+    @classmethod
+    def from_task_doc(
+        cls,
+        task_doc: TaskDoc,
+        **kwargs,
+    ) -> list[Self]:
+        """Create a list of ML training documents from the ionic steps in a TaskDoc.
+        
+        Parameters
+        -----------
+        task_doc : TaskDoc
+        **kwargs
+            Any kwargs to pass to `from_structure`.
+        """
+        entries = []
+
+        for cr in task_doc.calcs_reversed[::-1]:
+            nion = len(cr.output.ionic_steps)
+
+            for iion, ionic_step in enumerate(cr.output.ionic_steps):
+
+                structure = Structure.from_dict(ionic_step.structure.as_dict())
+                # these are fields that should only be set on the final frame of a calculation
+                # also patch in magmoms because of how Calculation works
+                last_step_kwargs = {}
+                if iion == nion - 1:
+                    if (magmom := cr.output.structure.site_properties.get("magmom")):
+                        structure.add_site_property("magmom",magmom)
+                    last_step_kwargs["bandgap"] = cr.output.bandgap
+                    if bader_analysis := cr.bader:
+                        for bk in (
+                            "charge",
+                            "magmom",
+                        ):
+                            last_step_kwargs[f"bader_{bk}s"] = bader_analysis[bk]
+
+                if (_st := ionic_step.stress) is not None:
+                    st = np.array(_st)
+                    if np.allclose(st,st.T,rtol=1e-8):
+                        # Stress tensor is symmetric
+                        last_step_kwargs["stress"] = matrix_3x3_to_voigt(_st)
+                    else:
+                        # Stress tensor is non-symmetric
+                        last_step_kwargs["stress_matrix"] = _st
+
+                entries.append(
+                    cls.from_structure(
+                        meta_structure=structure,
+                        energy=ionic_step.e_0_energy,
+                        forces=ionic_step.forces,
+                        functional=cr.run_type,
+                        **last_step_kwargs,
+                        **kwargs,
+                    )
+                )
+        return entries
+
+
+class MatPESProvenanceDoc(BaseModel):
+    """Information regarding the origins of a MatPES structure."""
+
+    original_mp_id: MPID | str | None = Field(
+        None,
+        description="MP identifier corresponding to the Materials Project structure from which this entry was sourced from.",
+    )
+    materials_project_version: str | None = Field(
+        None,
+        description="The version of the Materials Project from which the struture was sourced.",
+    )
+    md_ensemble: str = Field(
+        None,
+        description="The molecular dynamics ensemble used to generate this structure.",
+    )
+    md_temperature: float | None = Field(
+        None,
+        description="If a float, the temperature in Kelvin at which MLMD was performed.",
+    )
+    md_pressure: float | None = Field(
+        None,
+        description="If a float, the pressure in atmosphere at which MLMD was performed.",
+    )
+    md_step: int | None = Field(
+        None,
+        description="The step in the MD simulation from which the structure was sampled.",
+    )
+    mlip_name: str | None = Field(
+        None, description="The name of the ML potential used to perform MLMD."
+    )
+
+
+class MatPESTrainDoc(MLTrainDoc):
+    """
+    Schema for VASP data in the Materials Potential Energy Surface (MatPES) effort.
+    
+    """
+
+    matpes_id: str | None = Field(None, description="MatPES identifier.")
+
+    formation_energy_per_atom: float | None = Field(
+        None,
+        description="The uncorrected formation enthalpy per atom at zero pressure and temperature.",
+    )
+    cohesive_energy_per_atom: float | None = Field(
+        None, description="The uncorrected cohesive energy per atom."
+    )
+
+    provenance: MatPESProvenanceDoc | None = Field(
+        None, description="Information about the provenance of the structure."
+    )
+
+    @property
+    def pressure(self) -> float | None:
+        """Return the pressure from the DFT stress tensor."""
+        if self.stress:
+            return sum(self.stress[:3]) / 3.
+
+    @property
+    def magmoms(self) -> list[float] | None:
+        """Retrieve on-site magnetic moments from the structure if they exist."""
+        if self.structure and (magmom := self.structure.site_properties.get("magmom")):
+            return magmom
