+++ emmet-core/tests/test_ml.py
@@ -1,11 +1,17 @@
 import pytest
 
 # from matcalc.utils import get_universal_calculator
+import numpy as np
 from pymatgen.core import Structure
 from pymatgen.util.testing import PymatgenTest
 
 from emmet.core.elasticity import BulkModulus, ElasticTensorDoc, ShearModulus
-from emmet.core.ml import MLDoc
+from emmet.core.math import matrix_3x3_to_voigt
+from emmet.core.ml import MLDoc, MatPESTrainDoc
+from emmet.core.tasks import TaskDoc
+
+from tests.conftest import get_test_object
+
 
 # if TYPE_CHECKING:
 #    from ase.calculators.calculator import Calculator
@@ -74,3 +80,37 @@ def test_ml_doc(calculator, prop_kwargs: dict) -> None:
         assert isinstance(
             actual, typ
         ), f"{key=} expected type={typ.__name__}, got {type(actual).__name__}"
+
+
+def test_matpes_doc_from_task_doc(test_dir):
+
+    task_doc = TaskDoc.from_directory(
+        test_dir / "vasp" / get_test_object("SiOptimizeDouble").folder
+    )
+    matpes_train_docs = MatPESTrainDoc.from_task_doc(task_doc)
+
+    assert len(matpes_train_docs) == sum(len(cr.output.ionic_steps) for cr in task_doc.calcs_reversed)
+        
+    ctr = 0
+    for cr in task_doc.calcs_reversed[::-1]:
+        for iistep, istep in enumerate(cr.output.ionic_steps):
+            assert matpes_train_docs[ctr].energy == pytest.approx(istep.e_0_energy)
+            assert np.allclose(
+                matpes_train_docs[ctr].forces, istep.forces
+            )
+
+            assert np.allclose(
+                matpes_train_docs[ctr].stress, matrix_3x3_to_voigt(istep.stress)
+            )
+
+            if iistep < len(cr.output.ionic_steps) - 1:
+                assert matpes_train_docs[ctr].bandgap is None
+                assert matpes_train_docs[ctr].structure.site_properties.get("magmom") is None
+            else:
+                assert matpes_train_docs[ctr].bandgap == pytest.approx(cr.output.bandgap)
+                assert np.allclose(
+                    matpes_train_docs[ctr].structure.site_properties.get("magmom"),
+                    cr.output.structure.site_properties.get("magmom")
+                )
+            
+            ctr += 1
